+++ apps/interfaces/public/feeds/urls.py
@@ -1,8 +1,6 @@
 from django.urls import path
 from interfaces.public.feeds import views
 
-app_name = "feeds"
-
 urlpatterns = [
     path("feed/", views.AllEntriesFeed(), name="feed"),
     path("<slug:stream_slug>/feed/", views.StreamFeed(), name="stream_feed"),
