+++ apps/interfaces/public/home/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path("", views.HomeView.as_view(), name="home"),
+]
