+++ apps/interfaces/public/maps/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path("maps/cluster", views.cluster_map, name="cluster_map"),
+]
