+++ apps/interfaces/public/search/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path("search/", views.SearchView.as_view(), name="search"),
+]
