+++ apps/interfaces/public/urls.py
@@ -1,11 +1,34 @@
 from django.urls import include, path
 
+app_name = "public"
+
 urlpatterns = [
-    path("", include("interfaces.public.feeds.urls", namespace="feeds")),
+    path("", include("interfaces.public.feeds.urls")),
+    path(
+        "",
+        include(
+            "interfaces.public.home.urls",
+        ),
+    ),
+    path("", include("interfaces.public.entry.urls")),
+    path("", include("interfaces.public.authors.urls")),
+    path(
+        "",
+        include(
+            "interfaces.public.search.urls",
+        ),
+    ),
+    path(
+        "",
+        include(
+            "interfaces.public.trips.urls",
+        ),
+    ),
+    path(
+        "",
+        include(
+            "interfaces.public.maps.urls",
+        ),
+    ),
+    path("", include("interfaces.public.streams.urls")),
 ]
-
-
-# Public urls are last so "slug-like" urls in plugins are not matched to the stream-list view.
-urlpatterns.append(
-    path("", include("public.urls", namespace="public")),
-)
